+++ pkg/tracing/main.go
@@ -2,7 +2,6 @@ package tracing
 
 import (
 	"fmt"
-	"io"
 	"net/http"
 	"regexp"
 	"strconv"
@@ -13,8 +12,8 @@ import (
 )
 
 type AzureSDKStatistics struct {
-	AzureApiDuration  *prometheus.HistogramVec
-	AzureApiRateLimit *prometheus.GaugeVec
+	AzureAPIDuration  *prometheus.HistogramVec
+	AzureAPIRateLimit *prometheus.GaugeVec
 	Transport         http.RoundTripper
 }
 
@@ -22,7 +21,7 @@ var subscriptionRegexp = regexp.MustCompile(`^(?i)/subscriptions/([^/]+)/?.*$`)
 
 func New(registry prometheus.Registerer, transport http.RoundTripper) *AzureSDKStatistics {
 	stats := &AzureSDKStatistics{}
-	stats.AzureApiDuration = prometheus.NewHistogramVec(
+	stats.AzureAPIDuration = prometheus.NewHistogramVec(
 		prometheus.HistogramOpts{
 			Name:    "azurerm_api_http_request_duration_seconds",
 			Help:    "A histogram of request latencies.",
@@ -31,9 +30,9 @@ func New(registry prometheus.Registerer, transport http.RoundTripper) *AzureSDKS
 		[]string{"method", "code"},
 	)
 
-	registry.MustRegister(stats.AzureApiDuration)
+	registry.MustRegister(stats.AzureAPIDuration)
 
-	stats.AzureApiRateLimit = prometheus.NewGaugeVec(
+	stats.AzureAPIRateLimit = prometheus.NewGaugeVec(
 		prometheus.GaugeOpts{
 			Name: "azurerm_api_ratelimit",
 			Help: "AzureRM API ratelimit",
@@ -46,27 +45,18 @@ func New(registry prometheus.Registerer, transport http.RoundTripper) *AzureSDKS
 		},
 	)
 
-	registry.MustRegister(stats.AzureApiRateLimit)
+	registry.MustRegister(stats.AzureAPIRateLimit)
 
-	stats.Transport = stats.scrapeRateLimits(promhttp.InstrumentRoundTripperDuration(stats.AzureApiDuration, transport))
+	stats.Transport = stats.scrapeRateLimits(promhttp.InstrumentRoundTripperDuration(stats.AzureAPIDuration, transport))
+
 	return stats
 }
 
 func (s *AzureSDKStatistics) scrapeRateLimits(next http.RoundTripper) promhttp.RoundTripperFunc {
 	return func(req *http.Request) (*http.Response, error) {
 		resp, err := next.RoundTrip(req)
-
-		if strings.HasSuffix(req.Host, "metrics.monitor.azure.com") {
-			defer resp.Body.Close()
-			body, _ := io.ReadAll(resp.Body)
-
-			_ = body
-
-			return nil, err
-		}
-
 		if err != nil {
-			return resp, err
+			return resp, err //nolint:wrapcheck
 		}
 
 		// get hostname (shorten it to 3 parts)
@@ -74,57 +64,69 @@ func (s *AzureSDKStatistics) scrapeRateLimits(next http.RoundTripper) promhttp.R
 		if hostnameParts := strings.Split(hostname, "."); len(hostnameParts) > 3 {
 			hostname = strings.Join(hostnameParts[len(hostnameParts)-3:], ".")
 		}
-		subscriptionId := ""
+
+		subscriptionID := ""
 		if matches := subscriptionRegexp.FindStringSubmatch(req.URL.RawPath); len(matches) >= 2 {
-			subscriptionId = strings.ToLower(matches[1])
+			subscriptionID = strings.ToLower(matches[1])
 		}
 
 		if strings.HasPrefix(req.URL.RawPath, "/providers/microsoft.resourcegraph/") {
-			s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-user-quota-remaining", "resourcegraph", "quota")
+			s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+				"x-ms-user-quota-remaining", "resourcegraph", "quota")
 		}
 
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-microsoft.consumption-tenant-requests", "consumption", "tenant-requests")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-microsoft.consumption-tenant-requests", "consumption", "tenant-requests")
 
 		// subscription rate limits
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-subscription-reads", "subscription", "reads")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-subscription-writes", "subscription", "writes")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-subscription-resource-requests", "subscription", "resourceRequests")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-subscription-resource-entities-read", "subscription", "resource-entities-read")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-subscription-reads", "subscription", "reads")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-subscription-writes", "subscription", "writes")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-subscription-resource-requests", "subscription", "resourceRequests")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-subscription-resource-entities-read", "subscription", "resource-entities-read")
 
 		// tenant rate limits
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-tenant-reads", "tenant", "reads")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-tenant-writes", "tenant", "writes")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-tenant-resource-requests", "tenant", "resource-requests")
-		s.collectAzureApiRateLimitMetric(resp, hostname, subscriptionId, "x-ms-ratelimit-remaining-tenant-resource-entities-read", "tenant", "resource-entities-read")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-tenant-reads", "tenant", "reads")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-tenant-writes", "tenant", "writes")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-tenant-resource-requests", "tenant", "resource-requests")
+		s.collectAzureAPIRateLimitMetric(resp, hostname, subscriptionID,
+			"x-ms-ratelimit-remaining-tenant-resource-entities-read", "tenant", "resource-entities-read")
 
 		return resp, nil
 	}
 }
 
-func (s *AzureSDKStatistics) collectAzureApiRateLimitMetric(r *http.Response, hostname, subscriptionId, headerName, scopeLabel, typeLabel string) {
+func (s *AzureSDKStatistics) collectAzureAPIRateLimitMetric(r *http.Response, hostname, subscriptionID, headerName, scopeLabel, typeLabel string) {
 	headerValue := r.Header.Get(headerName)
 
-	if v, err := strconv.ParseInt(headerValue, 10, 64); err == nil {
+	if value, err := strconv.ParseInt(headerValue, 10, 64); err == nil {
 		// single value
-		s.AzureApiRateLimit.With(prometheus.Labels{
+		s.AzureAPIRateLimit.With(prometheus.Labels{
 			"endpoint":       hostname,
-			"subscriptionID": subscriptionId,
+			"subscriptionID": subscriptionID,
 			"scope":          scopeLabel,
 			"type":           typeLabel,
-		}).Set(float64(v))
+		}).Set(float64(value))
 	} else if strings.Contains(headerValue, ":") {
 		// multi value (comma separated eg "QueriesPerHour:496,QueriesPerMin:37,QueriesPer10Sec:11")
 		for _, val := range strings.Split(headerValue, ",") {
 			if parts := strings.SplitN(val, ":", 2); len(parts) == 2 {
 				quotaName := parts[0]
 				quotaValue := parts[1]
-				if v, err := strconv.ParseInt(quotaValue, 10, 64); err == nil {
-					s.AzureApiRateLimit.With(prometheus.Labels{
+
+				if value, err = strconv.ParseInt(quotaValue, 10, 64); err == nil {
+					s.AzureAPIRateLimit.With(prometheus.Labels{
 						"endpoint":       hostname,
-						"subscriptionID": subscriptionId,
+						"subscriptionID": subscriptionID,
 						"scope":          scopeLabel,
 						"type":           fmt.Sprintf("%s.%s", typeLabel, quotaName),
-					}).Set(float64(v))
+					}).Set(float64(value))
 				}
 			}
 		}
